Path check fix

+++ Channels/Everybody_Votes_Channel/votes.py
@@ -515,7 +515,8 @@ def sign_file(name):
                 with open(path + "voting.bin", "rb") as source:
                     with open(path + "wc24dl/VOTING.BIN", "wb") as dest:
                         dest.write(source.read()[320:])
-                os.remove(path + "wc24dl.vff")
+                if os.path.exists(path + "wc24dl.vff"):
+                    os.remove(path + "wc24dl.vff")
                 subprocess.call(
                     [
                         "vfftool",
